Lockscreen charging info: show decimal in battery temperature

+++ packages/SettingsLib/src/com/android/settingslib/fuelgauge/BatteryStatus.java
@@ -52,7 +52,7 @@ public class BatteryStatus {
     public final int level;
     public final int plugged;
     public final int chargingStatus;
-    public final int temperature;
+    public final float temperature;
     public final int maxChargingCurrent;
     public final int maxChargingVoltage;
     public final int maxChargingWattage;
@@ -67,7 +67,7 @@ public class BatteryStatus {
 
     public BatteryStatus(int status, int level, int plugged, int chargingStatus,
             int maxChargingWattage, boolean present, int maxChargingCurrent, int maxChargingVoltage,
-            int temperature) {
+            float temperature) {
         this.status = status;
         this.level = level;
         this.plugged = plugged;
